Move role methods to Trait; Use LdapUser class instead of TestUser class for unit testing
