+++ Model/LdapUser.php
@@ -2,12 +2,12 @@
 
 namespace FR3D\LdapBundle\Model;
 
-use Symfony\Component\Security\Core\User\AdvancedUserInterface;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
  * Simple user object.
  */
-class LdapUser implements LdapUserInterface, AdvancedUserInterface
+class LdapUser implements LdapUserInterface, UserInterface
 {
     use UserRoleTrait;
 
@@ -24,16 +24,6 @@ class LdapUser implements LdapUserInterface, AdvancedUserInterface
     protected $username;
 
     /**
-     * @var string
-     */
-    protected $email;
-
-    /**
-     * @var bool
-     */
-    protected $enabled;
-
-    /**
      * Encrypted password. Must be persisted.
      *
      * @var string
@@ -41,31 +31,13 @@ class LdapUser implements LdapUserInterface, AdvancedUserInterface
     protected $password;
 
     /**
-     * @var bool
-     */
-    protected $locked;
-    /**
-     * @var bool
-     */
-    protected $expired;
-
-    /**
      * @var array
      */
     protected $roles;
 
-    /**
-     * @var bool
-     */
-    protected $credentialsExpired;
-
     public function __construct()
     {
-        $this->enabled = false;
-        $this->locked = false;
-        $this->expired = false;
         $this->roles = array();
-        $this->credentialsExpired = false;
     }
 
     /**
@@ -82,20 +54,6 @@ class LdapUser implements LdapUserInterface, AdvancedUserInterface
     public function setDn($dn)
     {
         $this->dn = $dn;
-    }
-
-    public function addRole($role)
-    {
-        $role = strtoupper($role);
-        if ($role === static::ROLE_DEFAULT) {
-            return $this;
-        }
-
-        if (!in_array($role, $this->roles, true)) {
-            $this->roles[] = $role;
-        }
-
-        return $this;
     }
 
     /**
@@ -116,90 +74,15 @@ class LdapUser implements LdapUserInterface, AdvancedUserInterface
         return null;
     }
 
-    public function getEmail()
-    {
-        return $this->email;
-    }
-
     /**
      * Gets the encrypted password normally.
      * Not available via ldap.
      *
-     * @return string empty
+     * @return string
      */
     public function getPassword()
     {
-        return '';
-    }
-
-    /**
-     * Returns the user roles.
-     *
-     * @return array The roles
-     */
-    public function getRoles()
-    {
-        $roles = $this->roles;
-
-        // we need to make sure to have at least one role
-        $roles[] = static::ROLE_DEFAULT;
-
-        return array_unique($roles);
-    }
-
-    /**
-     * Never use this to check if this user has access to anything!
-     *
-     * Use the SecurityContext, or an implementation of AccessDecisionManager
-     * instead, e.g.
-     *
-     *         $securityContext->isGranted('ROLE_USER');
-     *
-     * @param string $role
-     *
-     * @return bool
-     */
-    public function hasRole($role)
-    {
-        return in_array(strtoupper($role), $this->getRoles(), true);
-    }
-
-    public function isAccountNonExpired()
-    {
-        if (true === $this->expired) {
-            return false;
-        }
-
-        return true;
-    }
-
-    public function isAccountNonLocked()
-    {
-        return !$this->locked;
-    }
-
-    public function isCredentialsNonExpired()
-    {
-        if (true === $this->credentialsExpired) {
-            return false;
-        }
-
-        return true;
-    }
-
-    public function isEnabled()
-    {
-        return $this->enabled;
-    }
-
-    public function removeRole($role)
-    {
-        if (false !== $key = array_search(strtoupper($role), $this->roles, true)) {
-            unset($this->roles[$key]);
-            $this->roles = array_values($this->roles);
-        }
-
-        return $this;
+        return $this->password;
     }
 
     public function setUsername($username)
@@ -209,46 +92,9 @@ class LdapUser implements LdapUserInterface, AdvancedUserInterface
         return $this;
     }
 
-    /**
-     * @param bool $boolean
-     *
-     * @return User
-     */
-    public function setCredentialsExpired($boolean)
-    {
-        $this->credentialsExpired = $boolean;
-
-        return $this;
-    }
-
-    public function setEmail($email)
-    {
-        $this->email = $email;
-
-        return $this;
-    }
-
-    public function setEnabled($boolean)
-    {
-        $this->enabled = (Boolean) $boolean;
-
-        return $this;
-    }
-
     public function setPassword($password)
     {
         $this->password = $password;
-
-        return $this;
-    }
-
-    public function setRoles(array $roles)
-    {
-        $this->roles = array();
-
-        foreach ($roles as $role) {
-            $this->addRole($role);
-        }
 
         return $this;
     }
