+++ Model/LdapUserInterface.php
@@ -17,4 +17,46 @@ interface LdapUserInterface
      * @return string Distinguished Name
      */
     public function getDn();
+
+    /**
+     * Never use this to check if this user has access to anything!
+     *
+     * Use the SecurityContext, or an implementation of AccessDecisionManager
+     * instead, e.g.
+     *
+     *         $securityContext->isGranted('ROLE_USER');
+     *
+     * @param string $role
+     *
+     * @return bool
+     */
+    public function hasRole($role);
+
+    /**
+     * Sets the roles of the user.
+     * This overwrites any previous roles.
+     *
+     * @param array $roles
+     *
+     * @return self
+     */
+    public function setRoles(array $roles);
+
+    /**
+     * Adds a role to the user.
+     *
+     * @param string $role
+     *
+     * @return self
+     */
+    public function addRole($role);
+
+    /**
+     * Removes a role to the user.
+     *
+     * @param string $role
+     *
+     * @return self
+     */
+    public function removeRole($role);
 }
