+++ Tests/Driver/ZendLdapDriverMockedTest.php
@@ -3,7 +3,7 @@
 namespace FR3D\LdapBundle\Tests\Driver;
 
 use FR3D\LdapBundle\Driver\ZendLdapDriver;
-use FR3D\LdapBundle\Tests\TestUser;
+use FR3D\LdapBundle\Model\LdapUser;
 use Zend\Ldap\Exception\LdapException as ZendLdapException;
 use Zend\Ldap\Ldap;
 
@@ -64,7 +64,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'test_username';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -79,7 +79,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'test_username';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -94,7 +94,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $username = 'bad_username';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setUsername($username);
 
         $this->zend->expects($this->once())
@@ -109,7 +109,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $dn       = 'uid=test_username,ou=example,dc=com';
         $password = 'password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setDn($dn);
 
         $this->zend->expects($this->once())
@@ -124,7 +124,7 @@ class ZendLdapDriverMockedTest extends \PHPUnit_Framework_TestCase
     {
         $dn       = 'uid=test_username,ou=example,dc=com';
         $password = 'bad_password';
-        $user     = new TestUser();
+        $user     = new LdapUser();
         $user->setDn($dn);
 
         $this->zend->expects($this->once())
