--- Tests/TestUser.php
@@ -1,149 +1,0 @@
-<?php
-
-namespace FR3D\LdapBundle\Tests;
-
-use FR3D\LdapBundle\Model\LdapUserInterface;
-use Symfony\Component\Security\Core\User\AdvancedUserInterface;
-use Symfony\Component\Security\Core\User\UserInterface;
-
-class TestUser implements UserInterface, AdvancedUserInterface, LdapUserInterface
-{
-    private $username;
-    private $password;
-    private $enabled;
-    private $locked;
-    private $dn;
-    private $roles = array();
-
-    public function setUsername($username)
-    {
-        $this->username = $username;
-    }
-
-    public function getUsername()
-    {
-        return $this->username;
-    }
-
-    public function setPassword($password)
-    {
-        $this->password = $password;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getPassword()
-    {
-        return $this->password;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getSalt()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function eraseCredentials()
-    {
-        return null;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getRoles()
-    {
-        return $this->roles;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setEnabled($enabled)
-    {
-        $this->enabled = $enabled;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isEnabled()
-    {
-        return $this->enabled;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setLocked($locked)
-    {
-        $this->locked = $locked;
-    }
-
-    public function setRoles($roles)
-    {
-        $this->roles = $roles;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isAccountNonLocked()
-    {
-        return !$this->locked;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isAccountNonExpired()
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function isCredentialsNonExpired()
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function setDn($dn)
-    {
-        $this->dn = $dn;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function getDn()
-    {
-        return $this->dn;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function equals(UserInterface $user)
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    public function __toString()
-    {
-        return $this->getUsername();
-    }
-}
